+++ src/main/java/pl/smsapi/api/response/SendStatusResponse.java
@@ -1,6 +1,7 @@
 package pl.smsapi.api.response;
 
 import org.json.JSONArray;
+import org.json.JSONObject;
 
 public class SendStatusResponse extends StatusResponse {
 
@@ -9,6 +10,13 @@ public class SendStatusResponse extends StatusResponse {
     public SendStatusResponse(int count, int parts, JSONArray jsonArray) {
         super(count, jsonArray);
         this.parts = parts;
+    }
+
+    @Override
+    protected MessageResponse buildItem(JSONObject jsonObject) {
+        MessageResponse response = super.buildItem(jsonObject);
+        response.setToBeSentAtTimestamp(jsonObject.optInt("date_sent"));
+        return response;
     }
 
     public int getParts() {