+++ src/test/java/pl/smsapi/api/action/sms/StatusJsonMother.java
@@ -1,0 +1,25 @@
+package pl.smsapi.api.action.sms;
+
+public class StatusJsonMother {
+
+    public static String create() {
+        return
+            "{" +
+            "   \"count\":1," +
+            "   \"parts\":1," +
+            "   \"list\":[" +
+            "       {" +
+            "           \"id\":\"0f0f0f0f0f0f0f0f0f0f0f0f\"," +
+            "           \"points\":0.21," +
+            "           \"number\":\"48123123123\"," +
+            "           \"date_sent\":1717500698," +
+            "           \"submitted_number\":\"123123123\"," +
+            "           \"status\":\"QUEUE\"," +
+            "           \"error\":null," +
+            "           \"idx\":null," +
+            "           \"parts\":1" +
+            "       }" +
+            "   ]" +
+            "}";
+    }
+}